Finalized LevelSettings and LevelManager communication -- Started adding method comments

+++ Whack a Mole/Assets/Scripts/ScriptableObjects/LevelSettings.cs
@@ -16,8 +16,8 @@ public class LevelSettings : ScriptableObject
     public LevelDifficulty levelDifficulty;
     [Range(0, 180)] public int levelLengthInSeconds;
     [Range(0, 10)] public int scorePerGopher;
-    public int minSecondsToSpawn = 2;
-    public int maxSecondsToSpawn = 5;
+    public float minSecondsToSpawn = 2f;
+    public float maxSecondsToSpawn = 5f;
     public int minGopherLifetime, maxGopherLifetime = 0;
